backend/pyproject.toml
[tool.poetry]
name = "instant-chat"
version = "0.1.0"
description = ""
authors = ["Petya Salnikov <pet.salnikov@gomobile.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.2"
pydantic-settings = "^2.6.0"
uvicorn = "^0.32.0"
sqlalchemy = "^2.0.36"
asyncpg = "^0.29.0"
alembic = "^1.13.3"
psycopg2-binary = "^2.9.10"
redis = "^5.1.1"
websockets = "^13.1"
celery = {extras = ["redis"], version = "^5.4.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
